refactor: 돈 계산 로직에서 double 사용을 int로 수정

+++ src/main/java/domain/player/Money.java
@@ -19,8 +19,8 @@ public class Money {
         }
     }
 
-    public int multiply(double value) {
-        return (int) (this.amount * value);
+    public int multiply(int value) {
+        return this.amount * value;
     }
 
     public int getAmount() {
